+++ tests/errors/examples/error.reference.js
@@ -1,4 +1,4 @@
-async () => {
+module.exports = async () => {
   const result = unknownFunction();
   return result;
 };