+++ tests/errors/examples/error.undef.js
@@ -1,4 +1,4 @@
-async data => {
+module.exports = async data => {
   const result = data.unknownKey();
   return result;
 };