+++ tests/scripts/examples/arrow.js
@@ -1,1 +1,1 @@
-(a, b) => a + b;
+module.exports = (a, b) => a + b;