+++ tests/scripts/examples/function.js
@@ -1,4 +1,4 @@
-(function mul(a, b) {
+module.exports = function mul(a, b) {
   const result = a * b;
   return result;
-});
+};